prevent gcc 9.2 miscompiling Q_FOREACH
With gcc 9.2, Q_FOREACH only loops over the first element [1]. Fix
this by backporting [2]. That commit is for other reasons, but also
fixes this.

Original commit by Thiago Macieira <thiago.macieira@intel.com>:

It's possible to do without them, which probably makes the number of
supported compilers a lot bigger: they just need to support decltype()
or __typeof__.

That includes the Intel compiler. The old code was also apparently
working, but no one had realized the old workaround for some old version
was still in place.

The loop overhead is more or less the same. I have not done benchmarks,
but inspection of the generated assembly shows more or less the same
number of instructions.

Change-Id: I32d499c84a6ddd19d994b49f17a469acb5c3a3f1
Reviewed-by: Olivier Goffart <ogoffart@woboq.com>
Reviewed-by: Marc Mutz <marc.mutz@kdab.com>

[1] https://github.com/qt/qtbase/commit/c35a3f519007af44c3b364b9af86f6a336f6411b
[2] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=90617